consolidating both versions of the gitlab detector
